+++ src/index.tsx
@@ -68,6 +68,7 @@ import './renderers/Divider';
 import './renderers/Each';
 import './renderers/Flex';
 import './renderers/Form/index';
+import './renderers/Form/wrapControl';
 import './renderers/Form/Control';
 import './renderers/Form/Hidden';
 import './renderers/Form/InputText';
